refactor(router): rename push -> navigate
