+++ packages/router/__tests__/navigate.ts
@@ -1,7 +1,7 @@
-import {push} from '../src';
+import {navigate} from '../src';
 
 describe('@corpuscule/router', () => {
-  describe('push', () => {
+  describe('navigate', () => {
     let historyPushStateSpy: jasmine.Spy;
     let historyStateSpy: jasmine.Spy;
 
@@ -16,14 +16,14 @@ describe('@corpuscule/router', () => {
     });
 
     it('should push new state to history', () => {
-      push('/test');
+      navigate('/test');
       // tslint:disable-next-line:no-unbound-method
       expect(history.pushState).toHaveBeenCalledWith('/test', '', '/test');
       expect(historyStateSpy).toHaveBeenCalled();
     });
 
     it('should allow to add title to a page', () => {
-      push('/test', 'Test');
+      navigate('/test', 'Test');
       // tslint:disable-next-line:no-unbound-method
       expect(history.pushState).toHaveBeenCalledWith('/test', 'Test', '/test');
     });
@@ -38,7 +38,7 @@ describe('@corpuscule/router', () => {
 
       window.addEventListener('popstate', listener);
 
-      push('/test');
+      navigate('/test');
     });
   });
 });
