+++ packages/router/src/index.js
@@ -9,7 +9,7 @@ export {gearAdvanced, createRouterToken, isProviderAdvanced, outletAdvanced};
 export {default as createUrl} from 'universal-router/generateUrls';
 export {default as createRouter} from './createRouter';
 export {default as Link} from './Link';
-export {default as push} from './push';
+export {default as navigate} from './navigate';
 export * from './tokens/lifecycle';
 
 const defaultToken = createRouterToken();
